Separate Caporal and computeResult function
