+++ eval_teamwise/package.json
@@ -1,0 +1,4 @@
+{
+  "name" : "eval_teamwise",
+  "main" : "./eval_teamwise.js"
+}
