+++ eval_teamwise/scores.csv
@@ -1,0 +1,8 @@
+studentID,score
+2012A7PS001G,9
+2012A7PS005G,
+2011B1A7001G,8
+2012A7PS003G,4
+student1,5
+student2,2
+student3,3
