+++ eval_teamwise/team.csv
@@ -1,0 +1,8 @@
+studentID,teamID
+2012A7PS001G,1
+2012A7PS005G,1
+2011B1A7001G,4
+2012A7PS003G,4
+student1,3
+student2,4
+student3,3
